requirements.txt
fire
twine
click
pyyaml